Refactor dynamic sys module import in test_fibonacci.py
Renamed 'Fibonacci Numbers (fast).py' to 'fibonacci_fast.py' to remove space in name.
Updated 'test_fibonacci.py' to use a normal import instead of boilerplate `sys` logic.
