+++ Part-1/euler's project/Templates and Examples/fibonacci_fast.py
